+++ tests/fixtures/application_fixtures.py
@@ -1446,7 +1446,6 @@ def privacy_notice(db: Session) -> Generator:
             "name": "example privacy notice",
             "notice_key": "example_privacy_notice",
             "description": "a sample privacy notice configuration",
-            "origin": "privacy_notice_template_1",
             "regions": [
                 PrivacyNoticeRegion.us_ca,
                 PrivacyNoticeRegion.us_co,
@@ -1536,7 +1535,6 @@ def privacy_notice_us_co_third_party_sharing(db: Session) -> Generator:
             "name": "example privacy notice us_co third_party_sharing",
             "notice_key": "example_privacy_notice_us_co_third_party_sharing",
             "description": "a sample privacy notice configuration",
-            "origin": "privacy_notice_template_2",
             "regions": [PrivacyNoticeRegion.us_co],
             "consent_mechanism": ConsentMechanism.opt_in,
             "data_uses": ["third_party_sharing"],
@@ -1558,7 +1556,6 @@ def privacy_notice_us_co_provide_service_operations(db: Session) -> Generator:
             "name": "example privacy notice us_co provide.service.operations",
             "notice_key": "example_privacy_notice_us_co_provide.service.operations",
             "description": "a sample privacy notice configuration",
-            "origin": "privacy_notice_template_2",
             "regions": [PrivacyNoticeRegion.us_co],
             "consent_mechanism": ConsentMechanism.opt_in,
             "data_uses": ["essential.service.operations"],
@@ -1580,7 +1577,6 @@ def privacy_notice_eu_fr_provide_service_frontend_only(db: Session) -> Generator
             "name": "example privacy notice us_co provide.service.operations",
             "notice_key": "example_privacy_notice_us_co_provide.service.operations",
             "description": "a sample privacy notice configuration",
-            "origin": "privacy_notice_template_2",
             "regions": [PrivacyNoticeRegion.eu_fr],
             "consent_mechanism": ConsentMechanism.opt_in,
             "data_uses": ["essential.service"],
@@ -2036,12 +2032,8 @@ def privacy_preference_history(
             "anonymized_ip_address": "92.158.1.0",
             "email": "test@email.com",
             "method": "button",
-            "privacy_experience_config_history_id": privacy_experience_privacy_center.histories[
-                0
-            ].experience_config_history_id,
-            "privacy_experience_history_id": privacy_experience_privacy_center.histories[
-                0
-            ].id,
+            "privacy_experience_config_history_id": privacy_experience_privacy_center.experience_config.experience_config_history_id,
+            "privacy_experience_id": privacy_experience_privacy_center.id,
             "preference": "opt_out",
             "privacy_notice_history_id": privacy_notice_history.id,
             "provided_identity_id": provided_identity.id,
@@ -2116,18 +2108,12 @@ def privacy_experience_privacy_center(
         db=db,
         data={
             "component": ComponentType.privacy_center,
-            "disabled": True,
             "region": PrivacyNoticeRegion.us_co,
             "experience_config_id": experience_config_privacy_center.id,
-            "experience_config_history_id": experience_config_privacy_center.histories[
-                0
-            ].id,
         },
     )
 
     yield privacy_experience
-    for history in privacy_experience.histories:
-        history.delete(db)
     privacy_experience.delete(db)
 
 
@@ -2165,12 +2151,8 @@ def privacy_experience_overlay(db: Session, experience_config_overlay) -> Genera
             "component": ComponentType.overlay,
             "region": PrivacyNoticeRegion.us_ca,
             "experience_config_id": experience_config_overlay.id,
-            "experience_config_history_id": experience_config_overlay.histories[0].id,
-            "disabled": False,
         },
     )
 
     yield privacy_experience
-    for history in privacy_experience.histories:
-        history.delete(db)
     privacy_experience.delete(db)