+++ tests/ops/service/connectors/test_fides_connector.py
@@ -7,7 +7,7 @@ from httpx import Client
 from fides.api.graph.traversal import TraversalNode
 from fides.api.models.connectionconfig import ConnectionConfig, ConnectionTestStatus
 from fides.api.models.datasetconfig import DatasetConfig
-from fides.api.models.policy import ActionType, Policy
+from fides.api.models.policy import Policy
 from fides.api.models.privacy_request import PrivacyRequest, PrivacyRequestStatus
 from fides.api.service.connectors.fides.fides_client import FidesClient
 from fides.api.service.connectors.fides_connector import (